# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "同桌小程序",
        "title": "Deskmat",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "qianren",
            "email": "1911401642@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "119.3.2.168:4016",
    "basePath": "/api/v1/",
    "paths": {
...
        "/dailyrecord/card": {
            "get": {
                "description": "点击查看同桌名片",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dailyrecord"
                ],
                "summary": "查看同桌名片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "{\"error_code\":\"00001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}"
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录"
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 服务器错误"
                    }
                }
            }
        },
...
        "/dailyrecord/message": {
            "get": {
                "description": "查看本次同桌消息记录，包含时间戳",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dailyrecord"
                ],
                "summary": "浏览消息记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "{\"error_code\":\"00001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}"
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录"
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败"
                    }
                }
            }
        },
...
        "/square/tag": {
            "post": {
                "description": "\"在同桌广场中搜索标签返回对应名片\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "square"
                ],
                "summary": "搜索标签",
                "parameters": [
                    {
                        "description": "输入要搜索的标签",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Tag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Card"
                            }
                        }
                    },
                    "401": {
                        "description": "Lack Param Or Param Not Satisfiable."
                    },
                    "404": {
                        "description": "搜索失败"
                    }
                }
            }
        },
...
        "/user/infor": {
            "post": {
                "description": "\"获取用户的基本信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户界面",
                "parameters": [
                    {
                        "description": "用户学号",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    },
                    "401": {
                        "description": "身份验证失败"
                    },
                    "404": {
                        "description": "获取基本信息失败"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Card": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "college": {
                    "description": "学院",
                    "type": "string"
                },
                "declaration": {
                    "type": "string"
                },
                "grade": {
                    "description": "年级",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "infor": {
                    "type": "string"
                },
                "major": {
                    "description": "专业",
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                },
                "status": {
                    "description": "记录是否有同桌",
                    "type": "string"
                },
                "tag1": {
                    "type": "string"
                },
                "tag2": {
                    "type": "string"
                },
                "tag3": {
                    "type": "string"
                },
                "tag4": {
                    "type": "string"
                },
                "tag5": {
                    "type": "string"
                },
                "users_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Id": {
            "type": "object",
            "properties": {
                "users_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Tag": {
            "type": "object",
            "properties": {
                "tag": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "college": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                }
            }
        },
...
    }
}